feat: add runtime loading progress tracking
There was an issue with resources getting downloaded twice, once when
initiating (resource preloading phase) and then when the application is
run with the run btn click.

This issue appears to be mitigated when the console.log override isn't
being reassigned to the original console.log.

I have no clue what is going on here 🤷
Skill issues detected 🤡